Feature

+++ worker/narreddit.py
@@ -3,7 +3,7 @@ from scraper import Scraper
 from elevenlabs_tts import ElevenlabsTTS
 from google_tts import GoogleTTS
 from videoGen import VideoGenerator
-from forcedAligner import ForcedAligner
+from aeneas_aligner import AeneasAligner
 import os
 from gpt import GPT
 
@@ -13,7 +13,7 @@ class NarReddit:
         self.scraper = Scraper(env)
         self.googleTTS = GoogleTTS()
         self.elevenlabsTTS = ElevenlabsTTS(env)
-        self.forcedAligner = ForcedAligner(env['GENTLE_URL'])
+        self.aeneasAligner = AeneasAligner()
         self.videoGen = VideoGenerator(env)
         self.gpt = GPT(env)
 
@@ -39,11 +39,12 @@ class NarReddit:
         print(f"Created audio file: {audioFile}")
         return audioFile
 
-    def createSubtitles(self, editedPost, audioFile, filePrefix):
+    def createSubtitles(self, editedPost, audioFile, filePrefix, language):
         subtitlesPath = os.path.join(
             'shared', 'tts-audio-files', f'subtitles-{filePrefix}.srt')
-        subtitleText = self.gpt.getSubtitles(editedPost)
-        self.forcedAligner.align(audioFile, subtitleText, subtitlesPath)
+        subtitleText = editedPost
+        self.aeneasAligner.align(
+            audioFile, subtitleText, subtitlesPath, language)
         return subtitlesPath
 
     def generateVideo(self, audioFile, subtitlesPath, params, language, filePrefix):
@@ -59,28 +60,34 @@ class NarReddit:
         return videoFile
 
     def createVideo(self, params):
-        filePrefix = params['DOC_ID']
-        postTitle, postContent = self.scrapePost(params)
-        languages = params['LANGUAGES'].lower().split(',')
-        gender = self.gpt.getGender(postContent)
-        videos = []
-        ttsEngine = params['TTS_ENGINE'].upper()
+        try:
+            filePrefix = params['DOC_ID']
+            postTitle, postContent = self.scrapePost(params)
+            languages = params['LANGUAGES'].lower().split(',')
+            if self.gpt.moderationCheckPassed(postContent) == False:
+                raise Exception("Post failed moderation check")
+            gender = self.gpt.getGender(postContent)
+            videos = []
+            ttsEngine = params['TTS_ENGINE'].upper()
 
-        for language in languages:
-            editedPost = self.gpt.expandAcronymsAndAbbreviations(
-                postContent, language)
-            audioFile = self.generateAudio(
-                editedPost, gender, language, filePrefix, ttsEngine)
+            for language in languages:
+                editedPost = self.gpt.expandAcronymsAndAbbreviations(
+                    postContent, language)
+                audioFile = self.generateAudio(
+                    editedPost, gender, language, filePrefix, ttsEngine)
 
-            if params['SUBTITLES'] == True and language == 'english':
-                subtitlesPath = self.createSubtitles(
-                    editedPost, audioFile, filePrefix)
-            else:
-                subtitlesPath = None
+                if params['SUBTITLES'] == True:
+                    subtitlesPath = self.createSubtitles(
+                        editedPost, audioFile, filePrefix, language)
+                else:
+                    subtitlesPath = None
 
-            videos.append(self.generateVideo(
-                audioFile, subtitlesPath, params, language, filePrefix))
-            os.remove(audioFile)
-            if subtitlesPath:
-                os.remove(subtitlesPath)
-        return videos
+                videos.append(self.generateVideo(
+                    audioFile, subtitlesPath, params, language, filePrefix))
+                os.remove(audioFile)
+                if subtitlesPath:
+                    os.remove(subtitlesPath)
+            return videos
+
+        except Exception as e:
+            raise Exception(e)
